backend/requirements.txt
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-google-genai==2.1.5
langchain-mongodb==0.6.2
langchain-text-splitters==0.3.8
langsmith==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.13.2
pypdf==5.6.1
python-dotenv==1.1.0
python-multipart==0.0.20
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
Faker==37.4.0
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.7.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1